requirements.txt
streamlit==1.22.0
pandas==2.0.3
scikit-learn==1.3.0
snowflake-snowpark-python